Split out functionality into a seperate concern so that some functionality can be reused
